+++ RELEASES.md
@@ -295,3 +295,13 @@ Based on:
 - [typescript v3.0.0-alpha] .
 ### Releases
 - [NPM v3.0.0-alpha] https://www.npmjs.com/package/@speakeasy-sdks/speakeasy-bar/v/3.0.0-alpha - .
+
+## 2024-07-02 14:39:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.326.0 (2.359.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.0.0-alpha.1] .
+### Releases
+- [NPM v3.0.0-alpha.1] https://www.npmjs.com/package/@speakeasy-sdks/speakeasy-bar/v/3.0.0-alpha.1 - .